Revert "ACT-4202"
