+++ modules/activiti-cxf/pom.xml
@@ -64,13 +64,6 @@
           </execution>
         </executions>
       </plugin>
-      <plugin>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>2.7.1</version>
-        <configuration>
-         <forkMode>always</forkMode>
-        </configuration>
-      </plugin>
     </plugins>
     <pluginManagement>
       <plugins>
