+++ modules/activiti-cxf/src/main/java/org/activiti/engine/impl/webservice/CxfWebServiceClient.java
@@ -18,12 +18,8 @@ import java.util.Arrays;
 import java.util.Enumeration;
 
 import org.activiti.engine.ActivitiException;
-import org.activiti.engine.delegate.BpmnError;
 import org.apache.cxf.endpoint.Client;
-import org.apache.cxf.interceptor.Fault;
 import org.apache.cxf.jaxws.endpoint.dynamic.JaxWsDynamicClientFactory;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 
 /**
@@ -33,8 +29,6 @@ import org.slf4j.LoggerFactory;
  */
 public class CxfWebServiceClient implements SyncWebServiceClient {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(CxfWebServiceClient.class);
-    
   protected Client client;
   
   public CxfWebServiceClient(String wsdl) {
@@ -64,15 +58,6 @@ public class CxfWebServiceClient implements SyncWebServiceClient {
    * {@inheritDoc}}
    */
   public Object[] send(String methodName, Object[] arguments) throws Exception {
-    try {
-       return client.invoke(methodName, arguments);
-    } catch (Fault e) {
-       LOGGER.debug("Technical error calling WS", e);
-       throw new ActivitiException(e.getMessage(), e);
-    } catch (Exception e) {
-       // Other exceptions should be associated to business fault defined in the service WSDL
-       LOGGER.debug("Business error calling WS", e);
-       throw new BpmnError(e.getClass().getName(), e.getMessage());
-    }
+    return client.invoke(methodName, arguments);
   }
 }
