--- modules/activiti-cxf/src/test/java/org/activiti/engine/impl/webservice/MaxValueReachedFault.java
@@ -1,27 +1,0 @@
-/* Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.activiti.engine.impl.webservice;
-
-import javax.xml.ws.WebFault;
-
-/**
- * An implementation of a fault returned by the counter WS
- *
- * @author Christophe DENEUX
- */
-@WebFault(name = "MaxValueReachedFault", targetNamespace = "http://impl.engine.activiti.org/test")
-public class MaxValueReachedFault extends Exception {
-
-    private static final long serialVersionUID = -8740025365942332353L;
-
-}
