+++ modules/activiti-cxf/src/test/java/org/activiti/engine/impl/webservice/WebServiceImportTest.java
@@ -12,6 +12,7 @@
  */
 package org.activiti.engine.impl.webservice;
 
+import org.activiti.engine.impl.test.PluggableActivitiTestCase;
 import org.activiti.engine.test.Deployment;
 
 /**
@@ -19,7 +20,7 @@ import org.activiti.engine.test.Deployment;
  *
  * @author Christophe DENEUX
  */
-public class WebServiceImportTest extends AbstratWebServiceTask {
+public class WebServiceImportTest extends PluggableActivitiTestCase {
 
     @Deployment
     public void testImport() throws Exception {
