+++ modules/activiti-cxf/src/test/java/org/activiti/engine/impl/webservice/WebServiceMockImpl.java
@@ -16,8 +16,6 @@ import java.util.Date;
 
 import javax.jws.WebService;
 
-import org.apache.cxf.common.i18n.UncheckedException;
-
 /**
  * An implementation of a Counter WS
  *
@@ -45,14 +43,8 @@ public class WebServiceMockImpl implements WebServiceMock {
   /**
    * {@inheritDoc}
    */
-  public void inc() throws MaxValueReachedFault {
-    if (this.count == 123456) {
-      throw new RuntimeException("A runtime exception not expected in the processing of the web-service");
-    } else if (this.count != Integer.MAX_VALUE) {
-      this.count++;
-    } else  {
-      throw new MaxValueReachedFault();
-    }
+  public void inc() {
+    this.count++;
   }
 
   /**
