--- modules/activiti-cxf/src/test/resources/org/activiti/engine/impl/webservice/WebServiceTaskTest.testFaultManagement.bpmn20.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<definitions id="definitions"
-  xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
-  xmlns:activiti="http://activiti.org/bpmn"
-  targetNamespace="org.activiti.enginge.impl.webservice"
-  xmlns:tns="org.activiti.enginge.impl.webservice"
-  xmlns:webservice="http://webservice.activiti.org/">
-
-  <import importType="http://schemas.xmlsoap.org/wsdl/"
-          location="http://localhost:63081/webservicemock?wsdl"
-          namespace="http://webservice.activiti.org/" />
-
-  <process id="webServiceInvocation">
-
-    <startEvent id="theStart" />
-
-    <sequenceFlow id="flow1" sourceRef="theStart" targetRef="webService" />
-
-    <serviceTask id="webService"
-                 name="Web service invocation"
-                 implementation="##WebService"
-                 operationRef="tns:incOperation" />
-                 
-    <boundaryEvent id="faultMaxValue" attachedToRef="webService">
-      <errorEventDefinition errorRef="org.activiti.engine.impl.test.MaxValueReachedFault_Exception" />
-    </boundaryEvent>
-
-    <sequenceFlow id="flow2" sourceRef="webService" targetRef="theEnd" />
-    <sequenceFlow id="flow2a" sourceRef="faultMaxValue" targetRef="theEndWithError" />
-
-    <endEvent id="theEnd" />
-    <endEvent id="theEndWithError" />
-    
-  </process>
-
-
-  <itemDefinition id="incRequestItem" structureRef="webservice:inc" /><!-- QName of input element --> <!-- NEEDED FOR THE ARGUMENTS -->
-  <itemDefinition id="incResponseItem" structureRef="webservice:incResponse" /><!-- QName of output element -->
-
-  <message id="incRequestMessage" itemRef="tns:incRequestItem" name="incRequestMessage" />
-  <message id="incResponseMessage" itemRef="tns:incResponseItem" name="incResponseMessage" />
-
-
-  <!-- Interface: implementationRef = QName of WSDL Port Type -->
-  <interface name="Webservice Interface"> <!-- NEEDED FOR THE PORT -->
-    <!-- Operation: implementationRef = QName of WSDL Operation -->
-    <operation id="incOperation" name="Increase Operation" implementationRef="webservice:inc"> <!-- NEEDED FOR THE OPERATION NAME -->
-      <inMessageRef>tns:incRequestMessage</inMessageRef>
-      <outMessageRef>tns:incResponseMessage</outMessageRef>
-    </operation>
-  </interface>
-
-
-</definitions>
