+++ modules/activiti-cxf/src/test/resources/org/activiti/engine/impl/webservice/counterSchema.xsd
@@ -1,107 +1,105 @@
-<?xml version="1.0" encoding="UTF-8"?><!-- Published by JAX-WS RI at http://jax-ws.dev.java.net. RI's version is JAX-WS RI 2.1.6 in JDK 6. -->
-<xs:schema xmlns:tns="http://webservice.impl.engine.activiti.org/" xmlns:xs="http://www.w3.org/2001/XMLSchema"
-   version="1.0" targetNamespace="http://webservice.impl.engine.activiti.org/">
+<?xml version="1.0" encoding="UTF-8"?><!-- Published by JAX-WS RI at http://jax-ws.dev.java.net. RI's version is JAX-WS RI 2.1.6 in JDK 6. --><xs:schema xmlns:tns="http://webservice.impl.engine.activiti.org/counter" xmlns:xs="http://www.w3.org/2001/XMLSchema" version="1.0" targetNamespace="http://webservice.impl.engine.activiti.org/counter">
 
-   <xs:element name="getCount" type="tns:getCount"></xs:element>
+<xs:element name="getCount" type="tns:getCount"></xs:element>
 
-   <xs:element name="getCountResponse" type="tns:getCountResponse"></xs:element>
+<xs:element name="getCountResponse" type="tns:getCountResponse"></xs:element>
 
-   <xs:element name="inc" type="tns:inc"></xs:element>
+<xs:element name="inc" type="tns:inc"></xs:element>
 
-   <xs:element name="incResponse" type="tns:incResponse"></xs:element>
+<xs:element name="incResponse" type="tns:incResponse"></xs:element>
 
-   <xs:element name="prettyPrintCount" type="tns:prettyPrintCount"></xs:element>
+<xs:element name="prettyPrintCount" type="tns:prettyPrintCount"></xs:element>
 
-   <xs:element name="prettyPrintCountResponse" type="tns:prettyPrintCountResponse"></xs:element>
+<xs:element name="prettyPrintCountResponse" type="tns:prettyPrintCountResponse"></xs:element>
 
-   <xs:element name="noNameResult" type="tns:noNameResult"></xs:element>
+    <xs:element name="noNameResult" type="tns:noNameResult"></xs:element>
 
-   <xs:element name="noNameResultResponse" type="tns:noNameResultResponse"></xs:element>
+<xs:element name="noNameResultResponse" type="tns:noNameResultResponse"></xs:element>
 
-   <xs:element name="reservedWordAsName" type="tns:reservedWordAsName"></xs:element>
+    <xs:element name="reservedWordAsName" type="tns:reservedWordAsName"></xs:element>
 
-   <xs:element name="reservedWordAsNameResponse" type="tns:reservedWordAsNameResponse"></xs:element>
+<xs:element name="reservedWordAsNameResponse" type="tns:reservedWordAsNameResponse"></xs:element>
 
-   <xs:element name="reset" type="tns:reset"></xs:element>
+<xs:element name="reset" type="tns:reset"></xs:element>
 
-   <xs:element name="resetResponse" type="tns:resetResponse"></xs:element>
+<xs:element name="resetResponse" type="tns:resetResponse"></xs:element>
 
-   <xs:element name="setTo" type="tns:setTo"></xs:element>
+<xs:element name="setTo" type="tns:setTo"></xs:element>
 
-   <xs:element name="setToResponse" type="tns:setToResponse"></xs:element>
+<xs:element name="setToResponse" type="tns:setToResponse"></xs:element>
 
-   <xs:complexType name="prettyPrintCount">
-      <xs:sequence>
-         <xs:element name="prefix" type="xs:string" minOccurs="0"></xs:element>
-         <xs:element name="suffix" type="xs:string" minOccurs="0"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+<xs:complexType name="prettyPrintCount">
+<xs:sequence>
+<xs:element name="prefix" type="xs:string" minOccurs="0"></xs:element>
+<xs:element name="suffix" type="xs:string" minOccurs="0"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="prettyPrintCountResponse">
-      <xs:sequence>
-         <xs:element name="prettyPrint" type="xs:string" minOccurs="0"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+<xs:complexType name="prettyPrintCountResponse">
+<xs:sequence>
+<xs:element name="prettyPrint" type="xs:string" minOccurs="0"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="reservedWordAsName">
-      <xs:sequence>
-         <xs:element name="prefix" type="xs:string" minOccurs="0"></xs:element>
-         <xs:element name="suffix" type="xs:string" minOccurs="0"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+    <xs:complexType name="reservedWordAsName">
+<xs:sequence>
+<xs:element name="prefix" type="xs:string" minOccurs="0"></xs:element>
+<xs:element name="suffix" type="xs:string" minOccurs="0"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="reservedWordAsNameResponse">
-      <xs:sequence>
-         <xs:element name="static" type="xs:string" minOccurs="0"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+<xs:complexType name="reservedWordAsNameResponse">
+<xs:sequence>
+<xs:element name="static" type="xs:string" minOccurs="0"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="noNameResult">
-      <xs:sequence>
-         <xs:element name="prefix" type="xs:string" minOccurs="0"></xs:element>
-         <xs:element name="suffix" type="xs:string" minOccurs="0"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+    <xs:complexType name="noNameResult">
+<xs:sequence>
+<xs:element name="prefix" type="xs:string" minOccurs="0"></xs:element>
+<xs:element name="suffix" type="xs:string" minOccurs="0"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="noNameResultResponse">
-      <xs:sequence>
-         <xs:element name="return" type="xs:string" minOccurs="0"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+<xs:complexType name="noNameResultResponse">
+<xs:sequence>
+<xs:element name="return" type="xs:string" minOccurs="0"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="inc">
-      <xs:sequence></xs:sequence>
-   </xs:complexType>
+<xs:complexType name="inc">
+<xs:sequence></xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="incResponse">
-      <xs:sequence></xs:sequence>
-   </xs:complexType>
+<xs:complexType name="incResponse">
+<xs:sequence></xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="getCount">
-      <xs:sequence></xs:sequence>
-   </xs:complexType>
+<xs:complexType name="getCount">
+<xs:sequence></xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="getCountResponse">
-      <xs:sequence>
-         <xs:element name="count" type="xs:int"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+<xs:complexType name="getCountResponse">
+<xs:sequence>
+<xs:element name="count" type="xs:int"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="setTo">
-      <xs:sequence>
-         <xs:element name="value" type="xs:int"></xs:element>
-      </xs:sequence>
-   </xs:complexType>
+<xs:complexType name="setTo">
+<xs:sequence>
+<xs:element name="value" type="xs:int"></xs:element>
+</xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="setToResponse">
-      <xs:sequence></xs:sequence>
-   </xs:complexType>
+<xs:complexType name="setToResponse">
+<xs:sequence></xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="reset">
-      <xs:sequence></xs:sequence>
-   </xs:complexType>
+<xs:complexType name="reset">
+<xs:sequence></xs:sequence>
+</xs:complexType>
 
-   <xs:complexType name="resetResponse">
-      <xs:sequence></xs:sequence>
-   </xs:complexType>
+<xs:complexType name="resetResponse">
+<xs:sequence></xs:sequence>
+</xs:complexType>
 </xs:schema>
