+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/behavior/WebServiceActivityBehavior.java
@@ -15,11 +15,9 @@ package org.activiti.engine.impl.bpmn.behavior;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.activiti.engine.delegate.BpmnError;
 import org.activiti.engine.impl.bpmn.data.AbstractDataAssociation;
 import org.activiti.engine.impl.bpmn.data.IOSpecification;
 import org.activiti.engine.impl.bpmn.data.ItemInstance;
-import org.activiti.engine.impl.bpmn.helper.ErrorPropagation;
 import org.activiti.engine.impl.bpmn.webservice.MessageInstance;
 import org.activiti.engine.impl.bpmn.webservice.Operation;
 import org.activiti.engine.impl.pvm.delegate.ActivityExecution;
@@ -62,53 +60,34 @@ public class WebServiceActivityBehavior extends AbstractBpmnActivityBehavior {
   public void execute(ActivityExecution execution) throws Exception {
     MessageInstance message;
     
-    try {
-       if (ioSpecification != null) {
-         this.ioSpecification.initialize(execution);
-         ItemInstance inputItem = (ItemInstance) execution.getVariable(this.ioSpecification.getFirstDataInputName());
-         message = new MessageInstance(this.operation.getInMessage(), inputItem);
-       } else {
-         message = this.operation.getInMessage().createInstance();
-       }
-    
-       execution.setVariable(CURRENT_MESSAGE, message);
-    
-       this.fillMessage(message, execution);
-    
-       MessageInstance receivedMessage = this.operation.sendMessage(message);
-
-       execution.setVariable(CURRENT_MESSAGE, receivedMessage);
-
-       if (ioSpecification != null) {
-         String firstDataOutputName = this.ioSpecification.getFirstDataOutputName();
-         if (firstDataOutputName != null) {
-           ItemInstance outputItem = (ItemInstance) execution.getVariable(firstDataOutputName);
-           outputItem.getStructureInstance().loadFrom(receivedMessage.getStructureInstance().toArray());
-         }
-       }
-    
-       this.returnMessage(receivedMessage, execution);
-    
-       execution.setVariable(CURRENT_MESSAGE, null);
-       leave(execution);
-    } catch (Exception exc) {
-
-       Throwable cause = exc;
-       BpmnError error = null;
-       while (cause != null) {
-          if (cause instanceof BpmnError) {
-             error = (BpmnError) cause;
-             break;
-          }
-          cause = cause.getCause();
-       }
-
-       if (error != null) {
-          ErrorPropagation.propagateError(error, execution);
-       } else {
-          throw exc;
-       }
+    if (ioSpecification != null) {
+      this.ioSpecification.initialize(execution);
+      ItemInstance inputItem = (ItemInstance) execution.getVariable(this.ioSpecification.getFirstDataInputName());
+      message = new MessageInstance(this.operation.getInMessage(), inputItem);
+    } else {
+      message = this.operation.getInMessage().createInstance();
     }
+    
+    execution.setVariable(CURRENT_MESSAGE, message);
+    
+    this.fillMessage(message, execution);
+    
+    MessageInstance receivedMessage = this.operation.sendMessage(message);
+
+    execution.setVariable(CURRENT_MESSAGE, receivedMessage);
+
+    if (ioSpecification != null) {
+      String firstDataOutputName = this.ioSpecification.getFirstDataOutputName();
+      if (firstDataOutputName != null) {
+        ItemInstance outputItem = (ItemInstance) execution.getVariable(firstDataOutputName);
+        outputItem.getStructureInstance().loadFrom(receivedMessage.getStructureInstance().toArray());
+      }
+    }
+    
+    this.returnMessage(receivedMessage, execution);
+    
+    execution.setVariable(CURRENT_MESSAGE, null);
+    leave(execution);
   }
   
   private void returnMessage(MessageInstance message, ActivityExecution execution) {
