+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/webservice/Operation.java
@@ -46,7 +46,7 @@ public class Operation {
     setInMessage(inMessage);
   }
   
-  public MessageInstance sendMessage(MessageInstance message) throws Exception {
+  public MessageInstance sendMessage(MessageInstance message) {
     return this.implementation.sendFor(message, this);
   }
