+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/webservice/OperationImplementation.java
@@ -36,5 +36,5 @@ public interface OperationImplementation {
    * @param operation the operation that is interested on sending the message
    * @return the resulting message
    */
-  MessageInstance sendFor(MessageInstance message, Operation operation) throws Exception;
+  MessageInstance sendFor(MessageInstance message, Operation operation);
 }
